core: fix every real bug the ASan/UBSan/clang-tidy CI job actually surfaced
The linux-native-asan and linux-native-clang-tidy jobs wired in earlier this
phase have never successfully compiled this codebase before -- reproduced
the exact same build locally (Docker, same image, same flags) to iterate
faster than a 40-90 minute CI round-trip per fix. Every one of these was
confirmed against CAC's own copy of the same code to separate genuinely
new FIC bugs from pre-existing ones this was simply the first strict-flags
compile to ever catch; all needed fixing regardless of origin to get from
"never compiled under these flags" to green.

Real, `--with-sanitizers=integer` UB (not just warnings):
- test/util/random.h, test/amount_tests.cpp: FIC's own MAX_MONEY =
  INT64_MAX (no fixed cap, unlike Bitcoin's small 21e6*COIN) turns
  `MAX_MONEY + 1` -- harmless upstream -- into real signed-overflow UB.
  Promote to uint64_t before the +1 in both.
- policy/feerate.cpp: CFeeRate's constructor multiplies fee*1000 before
  dividing by size; upstream's small MAX_MONEY made this never overflow,
  FIC's doesn't have that ceiling. Widened to __int128 (precedent already
  in util/fastrange.h) so large-but-valid inputs still compute correctly,
  only saturating true overflow.
- consensus/tx_verify.cpp: GetMinFee(size_t nBytes, ...) silently truncated
  nBytes to uint32_t before CFeeRate::GetFee() -- the whole reason it takes
  size_t is so callers can pass a pathological value and get a
  MoneyRange()-capped result, not silent truncation. Clamp explicitly
  instead.

Real thread-safety gaps clang's analysis had never been compiled against:
- wallet/staking.cpp's GetStakeWeight() (and the SelectCoinsForStaking() /
  AvailableCoinsForStaking() it calls) is reachable from the Qt GUI's
  network-weight display through 5 layers of interface wrappers
  (qt/bitcoingui.cpp -> WalletModel -> wallet/interfaces.cpp ->
  CWallet::GetStakeWeight() -> node/interfaces.cpp) without ever acquiring
  cs_wallet. Same shape for rpc/blockchain.cpp's GetPoSKernelPS() reading
  chainman.m_best_header without cs_main. Added the missing locks at the
  point closest to the actually-guarded reads, and the EXCLUSIVE_LOCKS_
  REQUIRED annotations needed to keep the analyzer honest about it going
  forward, rather than just silencing this one call path.
- net_processing.cpp: ProcessBlock/ProcessNetBlock (CAC's own extra layer
  over upstream's ProcessBlock) never propagated the !m_peer_mutex negative
  capability its own GetPeerRef/MaybePunishNodeForBlock calls require.
- wallet/rpc/spend.cpp: a `const CWallet& wallet{*pwallet}` local alias
  broke the analyzer's ability to match it against the LOCK(pwallet->
  cs_wallet) already taken on pwallet itself -- pass pwallet.get() instead
  of introducing an alias solely to route the same object through.

Everything else was dead/stale, all confirmed pre-existing in CAC: unused
be32dec()/pwallet field/consensusParams/transactionView/tx_bytes, a
zero-length CRPCCommand array (rpc/fees.cpp never registers anything --
FIC has no dynamic fee estimator), and two @param Doxygen comments
(V2Transport's version_in, setEncryptionStatus's status) documenting
parameters or names that don't exist in the actual signatures.

Confirmed clean: a full local rebuild under the exact ASan/UBSan/integer/
float-divide-by-zero + -Werror flags this CI job uses now compiles with
zero errors (every binary, including the Qt GUI and fuzz targets), and the
unit test suite ran clean for the ~20 minutes it takes to get through under
sanitizers before this verification pass's own time budget ended -- no
failures or sanitizer trips in anything that ran. The real CI job (180min
timeout) will be the authoritative full run.

+++ firstislamiccoin-core/src/test/util/random.h
@@ -69,7 +69,12 @@ static inline bool InsecureRandBool()
 
 static inline CAmount InsecureRandMoneyAmount()
 {
-    return static_cast<CAmount>(InsecureRandRange(MAX_MONEY + 1));
+    // FirstIslamicCoin: MAX_MONEY is INT64_MAX here (no fixed cap, unlike
+    // Bitcoin's small 21e6*COIN), so plain `MAX_MONEY + 1` overflows CAmount
+    // (int64_t) before InsecureRandRange ever sees it. Promote to uint64_t
+    // first -- InsecureRandRange's own parameter type -- so the +1 lands in
+    // range (2^63, representable) instead of wrapping a signed integer.
+    return static_cast<CAmount>(InsecureRandRange(static_cast<uint64_t>(MAX_MONEY) + 1));
 }
 
 #endif // BITCOIN_TEST_UTIL_RANDOM_H
